Add a `list_sensors()` RPC to get sensors
This RPC is based on version 0.17 of the microgrid API instead of the
current 0.15 version, so users can already use the new interface and
make migrating to 0.17 more straight forward.

Instead of having only one sensor class as done in the current client,
we introduce a Sensor hierarchy, with a specific type for each sensor
type. This allows grouping sensors more easily (like problematic
sensors), and also provides a more Pythonic interface. The same approach
will be done for component in future versions of this client.

To be compatible with 0.17, sensors also have a "fake" lifetime that is
not really retrieved from the microgrid (it is filled with a "eternal"
lifetime, so sensors are present always).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>