Add a `Sensor` class
This class will be used to list and retrieve sensors information from
the microgrid. The class is based on version 0.17 of the microgrid API
instead of the current 0.15 version, so users can already use the new
interface and make migrating to 0.17 more straight forward.

It is also implemented having in mind that sensor types will be removed
in future releases (and it is actually not being set properly in the
0.15 service), so sensor types are being ignored.

To be compatible with 0.17, sensors also have a "fake" lifetime that is
not really retrieved from the microgrid (it is filled with a "eternal"
lifetime, so sensors are present always).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>