Add the base `Component` class
This class will be used to combine classic inheritance (all components
will derive from it), a mix-in to provide common functionality of all
components, and also build "algebraic data types", so they can be used
in `match` statements and easily do exhaustion checks to make sure all
possible types of components are handled when needed.

Note it doesn't make sense to make this a `abc.ABC` because there are no
abstract methods.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>