Add `Battery` component and sub-classes/types
Battery components have an attached battery type. To encode this into
the Python type system, we create a sub-class per each battery type,
similar to what we do with components.

The special types `UnspecifiedBattery` and `UnrecognizedBattery` are
added to represent a battery with type `UNSPECIFIED` and a battery with
a battery type we don't recognize (this could happen if using a newer
server providing new battery types) respectively.

On top of that, we define a `BatteryTypes` type alias to make it easy
to type-hint function that want to return all known battery types as a
type union instead of using inheritance.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>