Add `EvCharger` component and sub-classes/types
Like batteries, EV charger components have an attached type. To encode
this into the Python type system, we create a sub-class per each type.

As with batteries, special types are using to represent a EV charger
with type `UNSPECIFIED` and with type we don't recognize, and define a
type alias to make it easy to type-hint function that want to return all
known EV charger types as a type union instead of using inheritance.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>