Add `Inverter` component and sub-classes/types
Like batteries and EV chargers, inverter components have an attached
type. To encode this into the Python type system, we create a sub-class
per each type.

Special types are also used to represent a inverter with type
`UNSPECIFIED` and with type we don't recognize, and we define a type
alias to make it easy to type-hint function that want to return all
known inverter types as a type union instead of using inheritance.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>