Add `ProblematicComponent` and sub-classes
Problematic components are components that can't be mapped to known
category types (or are `UNSPECIFIED`).  They also include components
with mismatched a category, i.e. a component with a particular known
category but that also has category-specific information that doesn't
match the specified category. For example if the category is `BATTERY`
but the category-specific information is for a `INVERTER`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>