Add parsing of components from protobuf messages
The `component_from_proto()` function parses protobuf `Component`
messages and create the appropriate and specific component type of our
component hierarchy.

During parsing, a few consistency checks are done in the protobuf
message. Solvable errors are recovered as much as possible. Major issues
are logged as warnings, and minor issues are logged as debug.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>