Use IDs from frequenz-client-common
This commit updates the microgrid client to use IDs from the
`frequenz-client-common` package instead of defining them in the
microgrid client itself. This change is part of a broader effort to
centralize common functionality and types across different clients.

We also bump some minimum dependency versions of some other dependencies
to be able to work with the new `frequenz-client-common` dependency.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>