Support testing unary-stream calls that return a coroutine
Originally testing unary-stream calls assumed a `GrpcStreamBroadcaster`
is used, which means a `Receiver` is returned synchronously. But not all
unary-stream calls adopt this pattern, and some might be async, so
returning a coroutine instead.

With this commit we inspect the return value and if it is a coroutine,
we await for it and assign the awaited value as the client result.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>