Remove the timeout from the gRPC stream call
The timeout counts for the whole duration of the call, not each
individual streamed message, so it always times out, as we expect this
stream to basically last forever.

On top of this, disconnecting the stream seems to trigger some bug in
the SDK, so we better avoid it for now.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>