Add pylint exception for too-many-branches
When converting a component from its protobuf representation, we need to
handle multiple component types, so we end up with multiple branches in
the function, but this is OK, so we just silence the pylint warning for
that specific case.

This is in preparation to the merge to v0.x.x branch, as only the new
version detects this case.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>